fixed some visual studio warnings
Mostly type-cast related warnings. Commit testing.

+++ Src/Graphics/ModelImageGenerator.cpp
@@ -59,7 +59,7 @@ PGE::Texture* ModelImageGenerator::generate(const PGE::String& model, float scal
         // Further optimized if in initialization step, requires Model shader to be loaded at that point.
         gfxRes->setCameraUniforms(cam);
 
-        position = position * texSize;
+        position = position * (float)texSize;
         mi->setPosition(PGE::Vector3f(position.x, position.y, 0.f));
         mi->setRotation(rotation);
         mi->setScale(PGE::Vector3f(scale * texSize));
